fix: address review findings on unified transaction table
- Keep value="" on the placeholder option (fastcore drops falsy attrs),
  so an empty selection submits "" instead of the label text
- Pin form enctype to urlencoded (FastHTML defaults to multipart)
- Pass page_size through to pagination instead of assuming 50
- Initialize categories in the review page DB-error fallback to avoid
  a NameError masking the error alert
- Share FakeTransaction via tests/fakes.py instead of two copies
