+++ tests/test_xss_hardening.py
@@ -1,28 +1,13 @@
 """Tests for XSS hardening across pages."""
 
 import tempfile
-from dataclasses import dataclass
 from pathlib import Path
 
 from fafycat.core.database import CategoryORM
-
+from fakes import FakeTransaction
 
 XSS_PAYLOAD = "<img src=x onerror=alert(1)>"
 XSS_ESCAPED = "&lt;img src=x onerror=alert(1)&gt;"
-
-
-@dataclass
-class _FakeTransaction:
-    """Minimal transaction-like object for rendering tests."""
-
-    id: str = "abc123"
-    date: str = "2025-06-15"
-    description: str = "Test Store"
-    amount: float = -10.0
-    actual_category: str | None = None
-    predicted_category: str | None = None
-    confidence: float | None = 0.75
-    is_reviewed: bool = False
 
 
 class TestUploadXSSEscaping:
@@ -79,7 +64,7 @@ class TestReviewPageXSS:
         from fafycat.web.components.transaction_table import render_table
 
         categories = [CategoryORM(name=XSS_PAYLOAD, type="spending", budget=0.0)]
-        transactions = [_FakeTransaction(predicted_category=XSS_PAYLOAD)]
+        transactions = [FakeTransaction(predicted_category=XSS_PAYLOAD)]
 
         html = render_table(transactions, categories)
